все писменные замечания по ПР
