+++ internal/storages/userDB.go
@@ -2,14 +2,14 @@ package storages
 
 import (
 	"context"
+	"github.com/Sur0vy/cows_health.git/errors"
+	"github.com/Sur0vy/cows_health.git/helpers"
+	"github.com/Sur0vy/cows_health.git/logger"
 	"time"
 
 	"github.com/doug-martin/goqu/v9"
 	_ "github.com/doug-martin/goqu/v9/dialect/postgres"
 
-	"github.com/Sur0vy/cows_health.git/internal/errors"
-	"github.com/Sur0vy/cows_health.git/internal/helpers"
-	"github.com/Sur0vy/cows_health.git/internal/logger"
 	"github.com/Sur0vy/cows_health.git/internal/models"
 	"github.com/jmoiron/sqlx"
 )
