+++ internal/usecase/cow/handle_test.go
@@ -4,9 +4,9 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
-	"github.com/Sur0vy/cows_health.git/internal/logger"
-	storageMock "github.com/Sur0vy/cows_health.git/internal/mocks"
 	"github.com/Sur0vy/cows_health.git/internal/models"
+	"github.com/Sur0vy/cows_health.git/logger"
+	"github.com/Sur0vy/cows_health.git/mocks"
 	"github.com/labstack/echo/v4"
 	"github.com/stretchr/testify/assert"
 	"io"
