+++ internal/usecase/farm/router.go
@@ -1,9 +1,9 @@
 package farm
 
 import (
+	"github.com/Sur0vy/cows_health.git/logger"
 	"github.com/labstack/echo/v4"
 
-	"github.com/Sur0vy/cows_health.git/internal/logger"
 	"github.com/Sur0vy/cows_health.git/internal/models"
 )
