+++ internal/usecase/user/handle_test.go
@@ -3,6 +3,9 @@ package user
 import (
 	"bytes"
 	"context"
+	"github.com/Sur0vy/cows_health.git/errors"
+	"github.com/Sur0vy/cows_health.git/logger"
+	"github.com/Sur0vy/cows_health.git/mocks"
 	"io"
 	"net/http"
 	"net/http/httptest"
@@ -11,9 +14,6 @@ import (
 	"github.com/labstack/echo/v4"
 	"github.com/stretchr/testify/assert"
 
-	"github.com/Sur0vy/cows_health.git/internal/errors"
-	"github.com/Sur0vy/cows_health.git/internal/logger"
-	storageMock "github.com/Sur0vy/cows_health.git/internal/mocks"
 	"github.com/Sur0vy/cows_health.git/internal/models"
 )
